Add customizable/higher limit for recently used
I set the Recently Used as initial view. However since it's limited to 10 entries and the XFCE menu categories can be larger, there can be some white space. This is an attempt to give users option to get rid of the white space and icrease the limit. Please consider (something like) this.