Remove extra spacing around profile picture
Now that the resize widget in the corner is gone, there is no
longer a need to add a 10px margin around the profile picture to
balance it visually. Instead, just use 6px padding between the
username and the profile picture to match the other 6px padding
in the menu.