Fix menu toggling in Wayland
In Wayland the menu always receives focus out events to hide,
no matter the method used (toggling panel button, clicking off
the window, etc). This fixes the handling of those events to
make the panel button toggle properly.

To make remote event toggling work in Wayland it is necessary to
switch back to a time-based method for ignoring remote events that
happen too closely to a focus out event.