Add dragging to resize menu (Issue #112)
This restores the edge resizing that was removed to support Wayland.
Manually handle resizing the window, instead of relying on Gtk. This
works in labwc for Wayland and Xfwm for X11.

Because it can't detect which side the panel button is on, all sides
can be dragged from. It then snaps the window back to where it was
after resizing is finished.